30 namespace and repository slug validations shall add to the name errors (#33)
* Move slug validation errors to errors of the slug base.

* Don't save the subject outside of the shared examples

The slug must not yet have been set in order to create a copy of the
subject with the exact same slug. Only after saving the duplicate, we
must run the validation of the subject (which is also run on save).
Otherwise the duplicate would result in the exact same database entry as
the subject.