Add url_path_method to compute the url_path
We need to lazily compute the `url_path` after the slug has been set.
Because the slug is set in a `before_validation` hook, we need to do the
same with the `url_path`, but still before `save` is called (because
of the validations).

In the backend, we can't know the slug before the validation, so we need to
provide a way to read it between initialization and validation.